refactor jwt creation to a common helper
